feat: add build script and fix cross-architecture compatibility
- Add build.bat for command-line builds
  - Supports VS2017/2019/2022 detection
  - Automates NuGet package restoration
  - Builds solution in Release mode
- Fix data corruption issues:
  - Replace size_t with uint32_t in model structures
  - Ensures consistent data size across different architectures
- Update .gitignore:
  - Add packages/ directory
  - Add *.exe files

+++ src/nya/NyaAttributes.h
@@ -90,7 +90,7 @@ public:
 
     /** @brief Index of a texture to use if HasTexture field is true
      */
-    size_t Texture;
+    uint32_t Texture;
 
     /** @brief Read structure from stream
      * @param stream Data stream
